--- src/components/login.jsx
@@ -1,18 +1,0 @@
-import {useAuth0} from "@auth0/auth0-react";
-import React, {useState} from "react";
-import {Button} from "@mui/material";
-
-const LoginButton = ({isDisabled}) => {
-  const {loginWithRedirect, isAuthenticated} = useAuth0();
-
-  const handleLogin = async () => {
-    await loginWithRedirect({
-      audience: "https://api.ii.co.uk/enrolled",
-      scope: "ii360:base"
-    });
-  }
-
-  return <Button disabled={isAuthenticated || isDisabled} variant="contained" onClick={() => handleLogin()}>Log In</Button>;
-};
-
-export default LoginButton;
